Add MouseTracker and placeholder components
The initial goal was to confirm my understanding of when you do and
don't need to bind state-changing methods.

While setting up the components for that, I happened upon a mention of
render props, so I added MouseTracker while following along with the
[React guide on render
props](https://reactjs.org/docs/render-props.html).